requirements.txt
annotated-types==0.7.0
anyio==4.9.0
click==8.1.8
colorama==0.4.6
diskcache==5.6.3
fastapi==0.115.12
h11==0.14.0
idna==3.10
Jinja2==3.1.6
llama_cpp_python==0.3.8
MarkupSafe==3.0.2
numpy==2.2.4
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dotenv==1.1.0
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
